Fixes for RISC OS build problems and compiler warnings.
git-svn-id: file:///home/survex-svn/survex/trunk@549 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/commands.c
@@ -1201,7 +1201,7 @@ require(void)
    fp = ftell(file.fh);
    while (1) {
       unsigned int have, want;
-      have = strtoul(p, &p, 10);
+      have = (unsigned int)strtoul(p, &p, 10);
       want = read_uint();
       if (have > want) break;
       if (have < want) {
@@ -1209,7 +1209,7 @@ require(void)
 	 char *v;
 	 /* find end of version number */
 	 while (isdigit(ch) || ch == '.') nextch();
-	 len = ftell(file.fh) - fp;
+	 len = (size_t)(ftell(file.fh) - fp);
 	 v = osmalloc(len + 1);
 	 fseek(file.fh, fp, SEEK_SET);
 	 ch = ch_old;
